server/requirements.txt
fastapi
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
motor
opencv-python-headless
numpy
fer
pydantic